commit 2

+++ src/main/java/com/xerofy/journalApplication/service/JournalEntryServices.java
@@ -4,6 +4,8 @@ import com.xerofy.journalApplication.entity.JournalEntry;
 import com.xerofy.journalApplication.entity.User;
 import com.xerofy.journalApplication.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,6 +15,7 @@ import java.util.Optional;
 
 @Component
 public class JournalEntryServices {
+   //private static final Logger logger= LoggerFactory.getLogger(JournalEntryServices.class);
 
     @Autowired
     private JournalEntryRepository journalEntryRepository;
@@ -28,7 +31,7 @@ public class JournalEntryServices {
             User user=userServices.findByUserName(userName);
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userServices.saveEntry(user);
+            userServices.saveUser(user);
         } catch (Exception e) {
             throw new RuntimeException("An error occured during saving the entry: " +e);
         }
@@ -42,11 +45,20 @@ public class JournalEntryServices {
     public Optional<JournalEntry> findById(ObjectId id){
         return journalEntryRepository.findById(id);
     }
+    @Transactional
+    public boolean deleteById(ObjectId id,String userName){
+        boolean removed=false;
+        try{
+            User user= userServices.findByUserName(userName);
+            removed=user.getJournalEntries().removeIf(X ->X.getId().equals(id));
+            if(removed){
+                userServices.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("An Error Accored while deleting ID "+e);
+        }
+        return removed;
 
-    public void deleteById(ObjectId id,String userName){
-        User user= userServices.findByUserName(userName);
-        user.getJournalEntries().removeIf(X ->X.getId().equals(id));
-        userServices.saveEntry(user);
-        journalEntryRepository.deleteById(id);
     }
 }
